rename the bin directory to scripts to ake it more obvious what it contains
