+++ scripts/convert_and_dispatch_genotypes.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 import csv
 import sys
 from os import remove
