+++ scripts/create_reagents_for_run.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 import os
 import sys
 import time
