+++ scripts/prepare_discard_plate.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 import sys
 from logging import FileHandler
 from egcg_core.app_logging import logging_default as log_cfg
