+++ tests/test_assign_to_workflow_stage.py
@@ -1,6 +1,6 @@
 from tests.test_common import fake_all_inputs, TestEPP
 from unittest.mock import Mock, patch
-from bin import assign_to_workflow_stage
+from scripts import assign_to_workflow_stage
 
 
 class TestAsignWorkflowStage(TestEPP):
@@ -31,14 +31,14 @@ class TestAsignWorkflowStage(TestEPP):
         self.epp2._lims = Mock()
 
     def test_assign(self):
-        with patch('bin.assign_to_workflow_stage.get_workflow_stage', return_value=Mock(uri='a_uri')) as p:
+        with patch('scripts.assign_to_workflow_stage.get_workflow_stage', return_value=Mock(uri='a_uri')) as p:
             self.epp._run()
             p.assert_called_with(self.epp.lims, self.epp.workflow_name, 'a_stage_name')
             exp_artifacts = ['a1', 'a2']
             assert sorted([a.id for a in self.epp.lims.route_artifacts.call_args[0][0]]) == exp_artifacts
             assert self.epp.lims.route_artifacts.call_args[1] == {'stage_uri': 'a_uri'}
 
-        with patch('bin.assign_to_workflow_stage.get_workflow_stage', return_value=Mock(uri='a_uri')) as p:
+        with patch('scripts.assign_to_workflow_stage.get_workflow_stage', return_value=Mock(uri='a_uri')) as p:
             self.epp2._run()
             p.assert_called_with(self.epp2.lims, self.epp2.workflow_name, 'a_stage_name')
             print(self.epp2.lims.route_artifacts.call_count)
